safari-600.7-branch/LayoutTests/fast/dom/nested-script-exceptions.html

<!DOCTYPE html>
<title>Exceptions in nested scripts</title>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<p>This test checks that exceptions in nested scripts are logged
properly. An exception named "Badness 10,000" should be logged properly.</p>

<script>document.write("<script>throw 'Badness 10,000'</scr" + "ipt>");</script>
