LayoutTests/fast/dom/nested-script-exceptions-expected.txt

CONSOLE MESSAGE: line 1: Badness 10,000
This test checks that exceptions in nested scripts are logged properly. An exception named "Badness 10,000" should be logged properly.
